reorganize to use an internal mpsc event queue & processing thread
This allows use from tokio applications & other async runtimes.

This also drops the use of the tracing-serde-* crates and moves that
functionality within itself. These changes would have required
deviating even further from upstream in incompatible ways. This also
should improve performance because we're going to be cloning fewer
things that we don't end up even using.

Breaking Changes

This adds a `finish` function to `TracingModality` that must be called
before the main thread exits to ensure that all tracing messages have
been flushed to Modality.

This also removes the `ModalitySubscriber` type that was previously
impossible to construct anyway and replaces it with
`ModalityLayer::into_subscriber()`.